代碼重構

+++ work_record/include/model/affected_type_dict_model.h
@@ -1,8 +1,9 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 struct AffectedTypeDict {
-    int id;
+    int64_t id;
     std::string affected;
     std::string comment;
 };
