+++ work_record/include/model/employee_dict_model.h
@@ -1,12 +1,12 @@
 #pragma once
 #include <string>
-#include <nlohmann/json.hpp>
+#include <cstdint>
 
 struct EmployeeDict {
-    int id;
+    int64_t id;
     std::string name;  // 对应数据库的name字段
     std::string employee_number;
-    int department_id;
+    int64_t department_id;
     std::string comment;
     std::string create_time;
     std::string department_name; // 关联查询时使用
