+++ work_record/include/model/file_record_model.h
@@ -1,7 +1,9 @@
 #pragma once
+#include <cstdint>
 #include <string>
+
 struct FileRecord {
-    int id;
+    int64_t id;
     std::string file_name;
     std::string file_path;
     std::string create_time;
