+++ work_record/include/model/issue_progress_dict_model.h
@@ -1,7 +1,9 @@
 #pragma once
 #include <string>
+#include <cstdint>
+
 struct IssueProgressDict {
-    int id;
+    int64_t id;
     std::string progress;
     std::string progress_class;
     std::string comment;
