+++ work_record/include/model/issue_record_model.h
@@ -1,21 +1,23 @@
 #pragma once
 #include <string>
+#include <cstdint>
+
 struct IssueRecord {
-    int id;
+    int64_t id;
     std::string issue_title;
     std::string reported_by;
     std::string description;
-    int progress_id;
+    int64_t progress_id;
     std::string progress_name;
     std::string progress_class;
-    int affected_id;
+    int64_t affected_id;
     std::string affected_name;
-    int source_type_id;
+    int64_t source_type_id;
     std::string source_type_name;
     std::string create_time;
     std::string update_time;
     std::string responsible_person;
-    int employee_id;
+    int64_t employee_id;
     std::string employee_name;
-    int department_id;
+    int64_t department_id;
 };
