+++ work_record/include/model/requirement_record_model.h
@@ -1,18 +1,20 @@
 #pragma once
 #include <string>
+#include <cstdint>
+
 struct RequirementRecord {
-    int id;
+    int64_t id;
     std::string title;
-    int requirement_status_id;
+    int64_t requirement_status_id;
     std::string requirement_status_name;
     std::string requirement_status_class;
-    int affected_id;
+    int64_t affected_id;
     std::string affected_name;
-    int source_type_id;
+    int64_t source_type_id;
     std::string source_type_name;
     std::string create_time;
     std::string update_time;
-    int employee_id;
+    int64_t employee_id;
     std::string employee_name;
-    int department_id;
+    int64_t department_id;
 };
