+++ work_record/include/model/requirement_status_dict_model.h
@@ -1,7 +1,9 @@
 #pragma once
 #include <string>
+#include <cstdint>
+
 struct RequirementStatusDict {
-    int id;
+    int64_t id;
     std::string status;
     std::string comment;
     std::string requirement_status_class;
