+++ work_record/include/model/source_type_dict_model.h
@@ -1,8 +1,9 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 struct SourceTypeDict {
-    int id;
+    int64_t id;
     std::string type;
     std::string comment;
 };
