+++ work_record/include/model/work_record_status_dict_model.h
@@ -1,7 +1,9 @@
 #pragma once
 #include <string>
+#include <cstdint>
+
 struct WorkRecordStatusDict {
-    int id;
+    int64_t id;
     std::string status_name;
     std::string status_class;
 };
